fix: docker-compose-local.yml to avoid build error by caching explicitly, chore: move from utils/* to util/auth/*, feat: S3 config
