+++ build.gradle
@@ -45,6 +45,9 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-validation' // test invalid api dto
 	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.11' // swagger
 	developmentOnly 'org.springframework.boot:spring-boot-devtools' // 핫 리로드 용도
+	implementation 'com.amazonaws:aws-java-sdk-s3:1.12.780'
+	implementation 'io.awspring.cloud:spring-cloud-aws-starter:3.1.1'
+	implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
 }
 
 tasks.named('test') {
