+++ src/main/java/com/ppu/ppu/auth/service/AuthService.java
@@ -6,8 +6,8 @@ import com.ppu.ppu.exception.domain.AuthException;
 import com.ppu.ppu.user.LoginType;
 import com.ppu.ppu.user.User;
 import com.ppu.ppu.user.UserService;
-import com.ppu.ppu.utils.JwtUtil;
-import com.ppu.ppu.utils.TokenIssueUtil;
+import com.ppu.ppu.utils.auth.JwtUtil;
+import com.ppu.ppu.utils.auth.TokenIssueUtil;
 import lombok.RequiredArgsConstructor;
 import org.mindrot.jbcrypt.BCrypt;
 import org.springframework.stereotype.Service;
