+++ src/main/java/com/ppu/ppu/utils/auth/kakao/KakaoAuthPurpose.java
@@ -1,4 +1,4 @@
-package com.ppu.ppu.utils.kakao;
+package com.ppu.ppu.utils.auth.kakao;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
