+++ src/main/java/com/ppu/ppu/utils/auth/kakao/KakaoUtil.java
@@ -1,5 +1,6 @@
-package com.ppu.ppu.utils.kakao;
+package com.ppu.ppu.utils.auth.kakao;
 
+import com.ppu.ppu.utils.auth.kakao.dto.*;
 import com.ppu.ppu.utils.kakao.dto.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
