+++ src/main/java/com/ppu/ppu/utils/auth/kakao/dto/KakaoUserProfileDto.java
@@ -1,4 +1,4 @@
-package com.ppu.ppu.utils.kakao.dto;
+package com.ppu.ppu.utils.auth.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
